* Fixed indent spaces to use white spaces instead of tab.

+++ tests/uHTTPNetworkInterfaceTests.cpp
@@ -20,33 +20,33 @@ using namespace uHTTP;
 
 BOOST_AUTO_TEST_CASE(NetworkInterfaceTests)
 {
-	NetworkInterfaceList ifList;
-	GetHostAddresses(ifList);
-	
-	int ifCnt = ifList.size();
-	BOOST_CHECK(0 < ifCnt);
+  NetworkInterfaceList ifList;
+  GetHostAddresses(ifList);
+  
+  size_t ifCnt = ifList.size();
+  BOOST_CHECK(0 < ifCnt);
     
 }
 
 BOOST_AUTO_TEST_CASE(SocketIPV6Tests)
 {
     /*
-	NetworkInterfaceList ifList;
-	GetHostAddresses(ifList);
-	
-	int ifCnt = ifList.size();
-	BOOST_CHECK(0 < ifCnt);
+  NetworkInterfaceList ifList;
+  GetHostAddresses(ifList);
+  
+  int ifCnt = ifList.size();
+  BOOST_CHECK(0 < ifCnt);
     
-	string ipv6if;
-	for (int n=0; n<ifCnt; n++) {
-		NetworkInterface *netif = ifList.getNetworkInterface(n);
-		const char *ifaddr = netif->getAddress();
-		if (IsIPv6Address(ifaddr) == false)
-			continue;
-		ipv6if = ifaddr;
-	}
+  string ipv6if;
+  for (int n=0; n<ifCnt; n++) {
+    NetworkInterface *netif = ifList.getNetworkInterface(n);
+    const char *ifaddr = netif->getAddress();
+    if (IsIPv6Address(ifaddr) == false)
+      continue;
+    ipv6if = ifaddr;
+  }
     
-	BOOST_CHECK(0 < ipv6if.size());
+  BOOST_CHECK(0 < ipv6if.size());
      */
 }
